F_T30603 move MapScreenshotter  to core
- adjust use statements
